removing deprecated tests no selection functions no more observing generator+conditions
