+++ tests/test_input.py
@@ -8,40 +8,6 @@ import numpy as np
 import pytest
 
 
-@pytest.mark.timeout(5)
-@pytest.mark.skip
-def test_parse_config(input_args):
-    """Checks if input config files are parsed correctly"""
-    simulations = ["all", "two_gal", "multi_gal", "group"]
-    for simulation in simulations:
-        args = input_args(simulation=simulation)
-        sys.path.append(os.getcwd())
-        btk_input = __import__("btk_input")
-        config_dict = btk_input.read_configfile(
-            args.configfile, args.simulation, args.verbose
-        )
-        assert set(config_dict.keys()) == {"user_input", "simulation"}, (
-            "config_dict must have only 'user_input', 'simulation'. Found"
-            f"{config_dict.keys()}"
-        )
-        set_keys = set(config_dict["simulation"].keys())
-        if simulation == "all":
-            assert set(simulations[1:]) == set_keys, (
-                "Invalid simulation "
-                "keys.Expected "
-                "{simulations[1:]} "
-                "found {set_keys}"
-            )
-        else:
-            assert {simulation} == set_keys, (
-                "Invalid "
-                f"simulation name. "
-                f"Expected {simulation}. "
-                f"Got {set_keys}"
-            )
-    pass
-
-
 @pytest.mark.skip
 def test_input_draw(input_args, match_images):
     """Tests that objects are drawn correctly when btk is run with input config
@@ -572,76 +538,3 @@ def basic_metric_two_gal_multi(output_name):
         err_msg="Did not get desired detection metrics summary",
     )
     pass
-
-
-@pytest.mark.skip
-def test_metrics(input_args):
-    """Btk measure is run for input config yaml file for different measure
-    functions and simulations. The measure outputs written to file are compared
-    to the values they are supposed to get.
-    """
-    simulations = [
-        "two_gal",
-    ]
-    for simulation in simulations:
-        command = ["python3", "btk_input.py", "--configfile", "tests/test-config.yaml"]
-        subprocess.call(command + ["--simulation", simulation])
-        args = input_args(simulation=simulation)
-        sys.path.append(os.getcwd())
-        btk_input = __import__("btk_input")
-        config_dict = btk_input.read_configfile(
-            args.configfile, args.simulation, args.verbose
-        )
-        user_config_dict = config_dict["user_input"]
-        output_path = os.path.join(
-            user_config_dict["output_dir"], user_config_dict["output_name"]
-        )
-        output_name = os.path.join(output_path, simulation + "_metrics_results.dill")
-        try:
-            basic_metric_two_gal(output_name)
-        except FileNotFoundError as e:
-            print("btk output files not found")
-            raise e
-        except ValueError as e:
-            print("btk output files were not in correct format")
-            raise e
-        finally:
-            delete_output_file(user_config_dict, simulation)
-            print("deleted files")
-        pass
-
-    # TEST MULTIPROCESSING
-    for simulation in simulations:
-        command = [
-            "python3",
-            "btk_input.py",
-            "--configfile",
-            "tests/test-config.yaml",
-            "--multiprocess",
-            "--cpus",
-            "4",
-        ]
-        subprocess.call(command + ["--simulation", simulation])
-        args = input_args(simulation=simulation)
-        sys.path.append(os.getcwd())
-        btk_input = __import__("btk_input")
-        config_dict = btk_input.read_configfile(
-            args.configfile, args.simulation, args.verbose
-        )
-        user_config_dict = config_dict["user_input"]
-        output_path = os.path.join(
-            user_config_dict["output_dir"], user_config_dict["output_name"]
-        )
-        output_name = os.path.join(output_path, simulation + "_metrics_results.dill")
-        try:
-            basic_metric_two_gal_multi(output_name)
-        except FileNotFoundError as e:
-            print("btk output files not found")
-            raise e
-        except ValueError as e:
-            print("btk output files were not in correct format")
-            raise e
-        finally:
-            delete_output_file(user_config_dict, simulation)
-            print("deleted files")
-        pass
